// Frey/app/clients/proto/frey.pb.ts
export interface PersonsCreateAccountPayload {
  email: string;
  password: string;
  title: string;
  city: string;
  phone: string;
  tenhouId: string;
  country: string;
}

export interface PersonsCreateAccountResponse {
  personId: number;
}
// ...
export interface PersonsFindByTenhouIdsPayload {
  ids: string[];
}

export interface PersonsFindByMajsoulIdsPayload {
  ids: MajsoulSearchEx[];
}

export interface MajsoulSearchEx {
  nickname: string;
  accountId: number;
}

export interface PersonsFindByTenhouIdsResponse {
  people: protoAtoms.PersonEx[];
}

export interface PersonsFindByTitlePayload {
  query: string;
}

export interface PersonsFindByTitleResponse {
  people: protoAtoms.Person[];
}
// ...
export const PersonsCreateAccountPayload = {
  /**
   * Serializes PersonsCreateAccountPayload to protobuf.
   */
  encode: function (msg: PartialDeep<PersonsCreateAccountPayload>): Uint8Array {
    return PersonsCreateAccountPayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes PersonsCreateAccountPayload from protobuf.
   */
  decode: function (bytes: ByteSource): PersonsCreateAccountPayload {
    return PersonsCreateAccountPayload._readMessage(
      PersonsCreateAccountPayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes PersonsCreateAccountPayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<PersonsCreateAccountPayload>,
  ): PersonsCreateAccountPayload {
    return {
      email: "",
      password: "",
      title: "",
      city: "",
      phone: "",
      tenhouId: "",
      country: "",
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<PersonsCreateAccountPayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.email) {
      writer.writeString(1, msg.email);
    }
    if (msg.password) {
      writer.writeString(2, msg.password);
    }
    if (msg.title) {
      writer.writeString(3, msg.title);
    }
    if (msg.city) {
      writer.writeString(4, msg.city);
    }
    if (msg.phone) {
      writer.writeString(5, msg.phone);
    }
    if (msg.tenhouId) {
      writer.writeString(6, msg.tenhouId);
    }
    if (msg.country) {
      writer.writeString(7, msg.country);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: PersonsCreateAccountPayload,
    reader: protoscript.BinaryReader,
  ): PersonsCreateAccountPayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.email = reader.readString();
          break;
        }
        case 2: {
          msg.password = reader.readString();
          break;
        }
        case 3: {
          msg.title = reader.readString();
          break;
        }
        case 4: {
          msg.city = reader.readString();
          break;
        }
        case 5: {
          msg.phone = reader.readString();
          break;
        }
        case 6: {
          msg.tenhouId = reader.readString();
          break;
        }
        case 7: {
          msg.country = reader.readString();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const PersonsCreateAccountResponse = {
  /**
   * Serializes PersonsCreateAccountResponse to protobuf.
   */
  encode: function (
    msg: PartialDeep<PersonsCreateAccountResponse>,
  ): Uint8Array {
    return PersonsCreateAccountResponse._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes PersonsCreateAccountResponse from protobuf.
   */
  decode: function (bytes: ByteSource): PersonsCreateAccountResponse {
    return PersonsCreateAccountResponse._readMessage(
      PersonsCreateAccountResponse.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes PersonsCreateAccountResponse with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<PersonsCreateAccountResponse>,
  ): PersonsCreateAccountResponse {
    return {
      personId: 0,
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<PersonsCreateAccountResponse>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.personId) {
      writer.writeInt32(1, msg.personId);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: PersonsCreateAccountResponse,
    reader: protoscript.BinaryReader,
  ): PersonsCreateAccountResponse {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.personId = reader.readInt32();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};
// ...
export const PersonsFindByTenhouIdsPayload = {
  /**
   * Serializes PersonsFindByTenhouIdsPayload to protobuf.
   */
  encode: function (
    msg: PartialDeep<PersonsFindByTenhouIdsPayload>,
  ): Uint8Array {
    return PersonsFindByTenhouIdsPayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes PersonsFindByTenhouIdsPayload from protobuf.
   */
  decode: function (bytes: ByteSource): PersonsFindByTenhouIdsPayload {
    return PersonsFindByTenhouIdsPayload._readMessage(
      PersonsFindByTenhouIdsPayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes PersonsFindByTenhouIdsPayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<PersonsFindByTenhouIdsPayload>,
  ): PersonsFindByTenhouIdsPayload {
    return {
      ids: [],
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<PersonsFindByTenhouIdsPayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.ids?.length) {
      writer.writeRepeatedString(1, msg.ids);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: PersonsFindByTenhouIdsPayload,
    reader: protoscript.BinaryReader,
  ): PersonsFindByTenhouIdsPayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.ids.push(reader.readString());
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const PersonsFindByMajsoulIdsPayload = {
  /**
   * Serializes PersonsFindByMajsoulIdsPayload to protobuf.
   */
  encode: function (
    msg: PartialDeep<PersonsFindByMajsoulIdsPayload>,
  ): Uint8Array {
    return PersonsFindByMajsoulIdsPayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes PersonsFindByMajsoulIdsPayload from protobuf.
   */
  decode: function (bytes: ByteSource): PersonsFindByMajsoulIdsPayload {
    return PersonsFindByMajsoulIdsPayload._readMessage(
      PersonsFindByMajsoulIdsPayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes PersonsFindByMajsoulIdsPayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<PersonsFindByMajsoulIdsPayload>,
  ): PersonsFindByMajsoulIdsPayload {
    return {
      ids: [],
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<PersonsFindByMajsoulIdsPayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.ids?.length) {
      writer.writeRepeatedMessage(
        1,
        msg.ids as any,
        MajsoulSearchEx._writeMessage,
      );
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: PersonsFindByMajsoulIdsPayload,
    reader: protoscript.BinaryReader,
  ): PersonsFindByMajsoulIdsPayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          const m = MajsoulSearchEx.initialize();
          reader.readMessage(m, MajsoulSearchEx._readMessage);
          msg.ids.push(m);
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};
// ...
export const PersonsFindByTenhouIdsResponse = {
  /**
   * Serializes PersonsFindByTenhouIdsResponse to protobuf.
   */
  encode: function (
    msg: PartialDeep<PersonsFindByTenhouIdsResponse>,
  ): Uint8Array {
    return PersonsFindByTenhouIdsResponse._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes PersonsFindByTenhouIdsResponse from protobuf.
   */
  decode: function (bytes: ByteSource): PersonsFindByTenhouIdsResponse {
    return PersonsFindByTenhouIdsResponse._readMessage(
      PersonsFindByTenhouIdsResponse.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes PersonsFindByTenhouIdsResponse with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<PersonsFindByTenhouIdsResponse>,
  ): PersonsFindByTenhouIdsResponse {
    return {
      people: [],
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<PersonsFindByTenhouIdsResponse>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.people?.length) {
      writer.writeRepeatedMessage(
        1,
        msg.people as any,
        protoAtoms.PersonEx._writeMessage,
      );
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: PersonsFindByTenhouIdsResponse,
    reader: protoscript.BinaryReader,
  ): PersonsFindByTenhouIdsResponse {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          const m = protoAtoms.PersonEx.initialize();
          reader.readMessage(m, protoAtoms.PersonEx._readMessage);
          msg.people.push(m);
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const PersonsFindByTitlePayload = {
  /**
   * Serializes PersonsFindByTitlePayload to protobuf.
   */
  encode: function (msg: PartialDeep<PersonsFindByTitlePayload>): Uint8Array {
    return PersonsFindByTitlePayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes PersonsFindByTitlePayload from protobuf.
   */
  decode: function (bytes: ByteSource): PersonsFindByTitlePayload {
    return PersonsFindByTitlePayload._readMessage(
      PersonsFindByTitlePayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes PersonsFindByTitlePayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<PersonsFindByTitlePayload>,
  ): PersonsFindByTitlePayload {
    return {
      query: "",
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<PersonsFindByTitlePayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.query) {
      writer.writeString(1, msg.query);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: PersonsFindByTitlePayload,
    reader: protoscript.BinaryReader,
  ): PersonsFindByTitlePayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.query = reader.readString();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};